interview-intelligence-evaluator/requirements.txt
# requirements.txt
streamlit
langchain
openai
chromadb
faiss-cpu
python-dotenv
PyPDF2
speechrecognition
whisper
